refactor: MVVMのパッケージ構造を作り、既存のActivityを.viewに入れた。
