+++ app/src/main/java/com/example/ttmatchlog/view/LoginActivity.kt
@@ -1,11 +1,8 @@
-package com.example.ttmatchlog
+package com.example.ttmatchlog.view
 
 import android.content.Intent
 import android.os.Bundle
-import androidx.activity.enableEdgeToEdge
 import androidx.appcompat.app.AppCompatActivity
-import androidx.core.view.ViewCompat
-import androidx.core.view.WindowInsetsCompat
 import com.example.ttmatchlog.databinding.ActivityLoginBinding
 
 class LoginActivity : AppCompatActivity() {
