+++ app/src/main/java/com/example/ttmatchlog/view/SigninActivity.kt
@@ -1,12 +1,8 @@
-package com.example.ttmatchlog
+package com.example.ttmatchlog.view
 
 import android.content.Intent
 import android.os.Bundle
-import android.util.Log
-import androidx.activity.enableEdgeToEdge
 import androidx.appcompat.app.AppCompatActivity
-import androidx.core.view.ViewCompat
-import androidx.core.view.WindowInsetsCompat
 import com.example.ttmatchlog.databinding.ActivitySigninBinding
 
 class SigninActivity : AppCompatActivity() {
