+++ app/src/main/res/layout/activity_signin.xml
@@ -8,7 +8,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:padding="30dp"
-    tools:context=".SigninActivity">
+    tools:context=".view.SigninActivity">
 
     <ImageView
         android:id="@+id/userIcon"
